requirements.txt
Django==5.2.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.7.0
python-decouple
celery==5.5.3
drf-nested-routers==0.94.2


Pillow==11.3.0
channels==4.3.1
daphne==4.2.1
whitenoise==6.10.0




django-debug-toolbar==6.0.0